fixed delay... still need to fix flash but everything else works

+++ led.c
@@ -57,13 +57,13 @@ void led_off(uint8_t ledIndex){
 
 void led_scan(){
 	// Scan the light across
-	for (uint8_t i = 0; i < 10; i++) {
+	for (int i = 0; i < 10; i++) {
 		led_on(i);
 		delay_1ms(BASE_FREQUENCY - currentSpeed * MULTIPLIER); 
 		led_off(i);
 	} 
 	// Scan the light back
-	for (uint8_t i = 9; i >= 0; i--) {
+	for (int i = 9; i >= 0; i--) {
 		led_on(i);
 		delay_1ms(BASE_FREQUENCY - currentSpeed * MULTIPLIER); 
 		led_off(i);
